Move icon to components dir
